move voxetta record button to allow merge
