+++ ui/src/VoxettaRecordButton.js
@@ -13,7 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License. */
 
-import {AudioRecorder} from '../AudioRecorder';
+import {AudioRecorder} from './AudioRecorder';
 import {LitElement, html, css} from 'lit-element';
 
 export class VoxettaRecordButton extends LitElement {
